*refactoring of ConvertIdentityNode.
